requirements.txt
enum34
pyserial
kivy
pyinstaller
nose
pyhamcrest
mockito-without-hardcoded-distribute-version
